feat(backends): rename LM Studio backend to OpenAI-compatible
The HTTP API was already OpenAI-spec compliant. This rename reflects that
any OpenAI-compatible server (LM Studio, Ollama, vLLM, etc.) works as a
drop-in backend.

- New: src/embeddings/openai_compat.rs, src/llm/openai_compat.rs
- Removed: src/embeddings/lmstudio.rs, src/llm/lmstudio.rs
- src/backends.rs re-exports OpenAiCompatEmbedder / OpenAiCompatLlm
- Config: lmstudio_base_url → api_base_url (alias kept for back-compat)
- Updated CLAUDE.md, SKILL.md, docs/ to reflect new naming
