+++ src/embeddings/mod.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
 
-pub mod lmstudio;
+pub mod openai_compat;
 
 /// The embedding vector dimension.
 /// `google/embeddinggemma-300m` and `BAAI/bge-base-en-v1.5` both output 768 dims.
