+++ src/llm/mod.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 use tokio::sync::mpsc;
 
-pub mod lmstudio;
+pub mod openai_compat;
 
 /// A streamed token from the LLM.
 pub type Token = String;
